Change how canvas is generated
The previous way of generating height and width for a IIIF canvas relied on images being stored locally, but ours are stored in S3. The way that this file generated the filename for `getimagesize` assumed local storage, and even if the file path could take the S3 storage adapter into account, the function would have to download the image to get the height and width. Instead, this change uses the image metadata that Omeka already stores to get the height and width.